requirements.txt
Pillow
praw
requests
prawcore
tweepy
schedule
python-dotenv